+++ app/States/DealFlow/Archived.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\States\DealFlow;
+
+class Archived extends DealFlowState
+{
+    public static string $name = 'archived';
+}